+++ src/layouts/BaseLayout.tsx
@@ -1,4 +1,5 @@
-import React from "react"
+import React from 'react'
+import { Header } from '../components/Header'
 
 interface BaseLayoutProps {
   children: React.ReactNode
@@ -7,6 +8,7 @@ interface BaseLayoutProps {
 export const BaseLayout = ({ children }: BaseLayoutProps) => {
   return (
     <>
+      <Header></Header>
       <header>header</header>
       {children}
       <footer>footer</footer>