+++ src/views/Lobby/index.tsx
@@ -1,7 +1,7 @@
-import React from "react"
-import { BaseLayout } from "../../layouts/BaseLayout"
-import { useSocketContext } from "../../contexts/SocketContext"
-import type { View } from "../../types/View"
+import React from 'react'
+import { BaseLayout } from '../../layouts/BaseLayout'
+import { useSocketContext } from '../../contexts/SocketContext'
+import type { View } from '../../types/View'
 
 interface LobbyProps {
   setView: (view: View) => void
@@ -22,7 +22,7 @@ export const Lobby = ({ setView }: LobbyProps) => {
         <h2 key={id}>{name}</h2>
       ))}
 
-      <button onClick={() => setView("EnterRoom")}>back</button>
+      <button onClick={() => setView('EnterRoom')}>back</button>
     </BaseLayout>
   )
 }